% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic}
\alias{historic}
\title{Download Historical Data on Brazilian Municipalities}
\usage{
historic(code = NULL, render = TRUE)
}
\arguments{
\item{code}{IBGE code. See \code{data("ibge_codes")}.}

\item{render}{In an interactive R session, whether to produce a HTML
page of the information that is returned by the function. This will
open in the RStudio viewer, if using RStudio.}
}
\description{
Download Historical Data on Brazilian Municipalities
}
